moved burning functions to ewutils, apply ghostbust and drunk effects when eating
