+++ client.py
@@ -482,7 +482,7 @@ async def on_ready():
 
 		asyncio.ensure_future(ewdistrict.capture_tick_loop(id_server = server.id))
 		asyncio.ensure_future(ewutils.bleed_tick_loop(id_server = server.id))
-		asyncio.ensure_future(ewstatuseffects.burn_tick_loop(id_server = server.id))
+		asyncio.ensure_future(ewutils.burn_tick_loop(id_server = server.id))
 		await ewtransport.init_transports(id_server = server.id)
 
 	try:
